Force to use bundled JRE for "installer" type
JRE is bundled to the Jira installer and Atlassian recommends to use it with Jira.
So, we don't need to manage Java separately if "installer" type is selected.

While defining JAVA_HOME, we should check the ['jira']['installer_type'] value
instead of ['java']['java_home'].
It helps to avoid collisions if node has already some ['java'] attributes assigned.